+++ Battlespace/Red_Enemy_Hard.h
@@ -12,7 +12,7 @@ public:
     Red_Enemy_Hard();
 
     int random_number = rand() % 550;
-    int red_life = 3;
+    float red_life = 3.00;
     int token = 0;
 
     void setEnemies_List(Linkedlist *enemies_list);
